rename folder name tat status to statuses
